DL-3004 SBT Update
