--- project/MicroServiceBuild.scala
@@ -1,39 +1,0 @@
-  import sbt._
-
-object MicroServiceBuild extends Build with MicroService {
-
-  val appName = "ated-subscription"
-
-  override lazy val appDependencies: Seq[ModuleID] = AppDependencies()
-}
-
-private object AppDependencies {
-  import play.sbt.PlayImport._
-  import play.core.PlayVersion
-
-  val compile: Seq[ModuleID] = Seq(
-    ws,
-    "uk.gov.hmrc" %% "bootstrap-play-26" % "1.3.0",
-    "uk.gov.hmrc" %% "domain"            % "5.6.0-play-26"
-  )
-
-  trait TestDependencies {
-    lazy val scope: String = "it,test"
-    val test : Seq[ModuleID]
-  }
-
-  object Test {
-    def apply() = new TestDependencies {
-      override lazy val test = Seq(
-        "org.scalatest"           %% "scalatest"          % "3.0.5"             % scope,
-        "org.scalatestplus.play"  %% "scalatestplus-play" % "3.1.2"             % scope,
-        "org.pegdown"              % "pegdown"            % "1.6.0"             % scope,
-        "com.typesafe.play"       %% "play-test"          % PlayVersion.current % scope,
-        "org.mockito"              % "mockito-core"       % "3.2.4"           % scope,
-        "com.github.tomakehurst"   % "wiremock-jre8"      % "2.23.2"            % IntegrationTest withSources()
-      )
-    }.test
-  }
-
-  def apply() = compile ++ Test()
-}
