+++ project/TestPhases.scala
@@ -1,10 +1,6 @@
-import sbt.Tests.{Group, SubProcess}
-import sbt.{ForkOptions, TestDefinition}
+import sbt._
 
 object TestPhases {
-
-  def oneForkedJvmPerTest(tests: Seq[TestDefinition]): Seq[Group] =
-    tests map {
-      test => Group(test.name, Seq(test), SubProcess(ForkOptions(runJVMOptions = Seq("-Dtest.name=" + test.name))))
-    }
+  lazy val TemplateTest = config("tt") extend Test
+  lazy val TemplateItTest = config("tit") extend IntegrationTest
 }
